==> WILO_FieldService_POC/ProductPrediction/07_LookUpPrediction/.ipynb_aml_checkpoints/02_LookUp_Train-checkpoint2020-5-22-11-0-51.ipynb <==
from azureml.core import Workspace, Experiment
from azureml.core.authentication import InteractiveLoginAuthentication
from azureml.train.estimator import Estimator
import json

interactive_auth = InteractiveLoginAuthentication(tenant_id="39288a38-ff19-432c-8011-1cd9d0dff445")
ws = Workspace(subscription_id="793146d9-d4dc-4a73-9728-76c4ffd0cc0d", resource_group="rg_dynamics_test", workspace_name="resdynml1test", auth=interactive_auth)

# load experiment cfg
with open("experiment_cfg.json", "r") as cfg_file:
    cfg = json.load(cfg_file)

## Training Script

%%writefile ./src/lookup.py

import numpy as np
import pandas as pd

class LookUpClassifier():
    
    def __init__(self):
        pass
    
    def fit(self, df):
        
        self.product_transform = list(set(x for l in df['ProductNrs'].apply(lambda x: x.split()).values.tolist() for x in l))
        
        symptoms_per_case_df = pd.DataFrame(df['Symptoms'].str.split(' ').tolist(), index=df['Job Card.JobCard Number']).stack().reset_index([0, 'Job Card.JobCard Number'])
        symptoms_per_case_df.columns = ['Job Card.JobCard Number', 'Symptom']

        prodnr_per_case_df = pd.DataFrame(df['ProductNrs'].str.split(' ').tolist(), index=df['Job Card.JobCard Number']).stack().reset_index([0, 'Job Card.JobCard Number'])
        prodnr_per_case_df.columns = ['Job Card.JobCard Number', 'ProductNr']

        df = pd.merge(symptoms_per_case_df, df, on='Job Card.JobCard Number', how='left')
        df = pd.merge(prodnr_per_case_df, df, on='Job Card.JobCard Number', how='left')

        df = df[['ProductId', 'Country', 'Symptom', 'ProductNr']].replace('', np.nan).dropna().reset_index(drop=True)

#         self.model = df.groupby(['ProductId', 'Country', 'Symptom'])#
        self.model = {}
        for i in range(len(df)):
            if not df['ProductId'][i] in self.model:
                self.model[df['ProductId'][i]] = {}
            if not df['Country'][i] in self.model[df['ProductId'][i]]:
                self.model[df['ProductId'][i]][df['Country'][i]] = {}
            if not df['Symptom'][i] in self.model[df['ProductId'][i]][df['Country'][i]]:
                self.model[df['ProductId'][i]][df['Country'][i]][df['Symptom'][i]] = []
            self.model[df['ProductId'][i]][df['Country'][i]][df['Symptom'][i]].append(df['ProductNr'][i])
    
    def predict(self, X):
        # X = [['<prodid>', '<country>', '<symptom1>, <symptom2>']]
        y = []
        for row in X:
            y_row = []
            for symptom in row[2].split(' '):
                if row[0] in self.model:
                    if row[1] in self.model[row[0]]:
                        if symptom in self.model[row[0]][row[1]]:
                            y_row += self.model[row[0]][row[1]][symptom]
            y.append(' '.join(map(str, list(set(y_row)))))
        return y
    
    def transform_products(self, y):
        # y = [ '<prod1> <prod2>', '<prod1> <prod3>' ]
        y_tr = np.zeros([len(y), len(self.product_transform)])
        for row in range(len(y)):
            for prod in y[row].split(' '):
                if prod in self.product_transform:
                    y_tr[row, self.product_transform.index(prod)] = 1
        return y_tr

%%writefile ./src/train.py

from azureml.core import Run

import os
import joblib
from argparse import ArgumentParser
from lookup import LookUpClassifier
from sklearn.metrics import recall_score, precision_score, hamming_loss, zero_one_loss, mean_absolute_error, mean_squared_error, r2_score

run = Run.get_context()

parser = ArgumentParser()
parser.add_argument('--input', dest='prepared_data')
args = parser.parse_args()

############################################################

# load data
if args.prepared_data:
    train_data = pd.read_csv(args.prepared_data + '/train_data.csv', sep=';', header=0)
    test_data = pd.read_csv(args.prepared_data + '/test_data.csv', sep=';', header=0)
else:
    train_data = run.input_datasets['train_data'].to_pandas_dataframe()
    test_data = run.input_datasets['test_data'].to_pandas_dataframe()
    
train_data = train_data.dropna().reset_index(drop=True)
test_data = test_data.dropna().reset_index(drop=True)
    
#################################################################

# train classifier
model = LookUpClassifier()
model.fit(train_data)

############################################################

X_test = test_data[['ProductId', 'Country', 'Symptoms']].values.tolist()
y_test = test_data['ProductNrs'].values.tolist() 

X_train = train_data[['ProductId', 'Country', 'Symptoms']].values.tolist()
y_train = train_data['ProductNrs'].values.tolist() 

############################################################

# evaluate test data
y_pred = model.predict(X_test)
y_pred_tr = model.transform_products(y_pred)
y_test_tr = model.transform_products(y_test)
run.log_table(
    'test_evaluation_classification',
    {
        'precision_macro': [precision_score(y_test_tr, y_pred_tr, average='macro')],
        'precision_samples': [precision_score(y_test_tr, y_pred_tr, average='samples')],
        'recall_macro': [recall_score(y_test_tr, y_pred_tr, average='macro')],
        'recall_samples': [recall_score(y_test_tr, y_pred_tr, average='samples')],
        'hamming_loss': [hamming_loss(y_test_tr, y_pred_tr)],
        'zero_one_loss': [zero_one_loss(y_test_tr, y_pred_tr)]
    }
)

# evaluate train data
y_pred = model.predict(X_train)
y_pred_tr = model.transform_products(y_pred)
y_train_tr = model.transform_products(y_train)
run.log_table(
    'train_evaluation_classification',
    {
        'precision_macro_train': [precision_score(y_train_tr, y_pred_tr, average='macro')],
        'precision_samples_train': [precision_score(y_train_tr, y_pred_tr, average='samples')],
        'recall_macro_train': [recall_score(y_train_tr, y_pred_tr, average='macro')],
        'recall_samples_train': [recall_score(y_train_tr, y_pred_tr, average='samples')],
        'hamming_loss_train': [hamming_loss(y_train_tr, y_pred_tr)],
        'zero_one_loss_train': [zero_one_loss(y_train_tr, y_pred_tr)]
    }
)

############################################################

# save model
os.makedirs('outputs', exist_ok=True)
joblib.dump(value=model, filename='outputs/model.pkl')

run.complete()

## Create Estimator

est = Estimator(entry_script='train.py', source_directory='src',
                inputs=[ws.datasets[cfg['train_dataset']].as_named_input('train_data'), 
                        ws.datasets[cfg['test_dataset']].as_named_input('test_data')   ],
                compute_target=ws.compute_targets['mlcompute'], environment_definition=ws.environments[cfg['env_name']])

## Run Experiment

exp = Experiment(ws, cfg['experiment_name'])
run = exp.submit(est)
run.wait_for_completion(show_output=True)
